+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 4.3.4
 -----
 
-- Fix crash on empty aggregate report comments (brakhane - PR #25)
+- Fix crash on empty aggregate report comments (brakhane - #25)
 - Add `strip_attachment_payloads` option to functions and 
-`--strip-attachment-payloads` option to the CLI
+`--strip-attachment-payloads` option to the CLI (#23)
 
 4.3.3
 -----