+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 -----
 
 - Fix crash on empty aggregate report comments (brakhane - #25)
+- Add SHA256 hashes of attachments to output
 - Add `strip_attachment_payloads` option to functions and 
 `--strip-attachment-payloads` option to the CLI (#23)
+- Remove `urllib3` from requirements to avoid conflict with `requests`
 
 4.3.3
 -----